Updated Nuxt dependency to stable version v3.0.0 and refactored project structure for Nuxt 3 (auto component import, folder structure, ...).
