+++ .npmrc
@@ -1,0 +1,3 @@
+shamefully-hoist=true
+auto-install-peers=false
+strict-peer-dependencies=false
