+++ package.json
@@ -10,7 +10,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.4.13",
-    "nuxt": "3.0.0-rc.13",
+    "nuxt": "3.0.0",
     "postcss": "^8.4.19",
     "tailwindcss": "^3.2.4"
   },
