--- src/pages/HomeView.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import WidgetComponent from '@/components/WidgetComponent.vue';
-</script>
-
-<template>
-  <main>
-    <WidgetComponent />
-  </main>
-</template>
